refactor(app-file-ops): extract trash and restore workflows
